Split source in modules. Generate API documentation. Add eslinter.
- Split the source code into modules inside `lib/`
- Generate browser ready distribution files into `dist/` (replaces the old `dsp.js` file)
- Add missing jsdoc documentation
- Generate API documentation into `docs/`
- Update the benchmarks code to use in node
- Update the examples to use the new `dsp.min.js` file
- Add eslint to make syntax coherent and prevent errors
- Add docs/README.md with the new install and development instructions
