+++ bench/bench.js
@@ -1,12 +1,12 @@
-function benchmark(func, loopCount) {
+module.exports = function benchmark(func, loopCount) {
   loopCount = loopCount || 10000;
 
   var start = Date.now();
 
-  for (var i = 0; i < loopCount; i++) { 
+  for (var i = 0; i < loopCount; i++) {
     func();
   }
 
   var end = Date.now();
   return end - start;
-}
+};
