+++ bench/dft.js
@@ -1,9 +1,13 @@
+/* global FREQUENCY */
+/* eslint-disable no-console */
+var benchmark = require("./bench");
+var dsp = require("..");
 var bufferSize = 2048;
 var sampleRate = 44100;
 var frequency = FREQUENCY || 440;
 
-var dft = new DFT(bufferSize, sampleRate);
-var osc = new Oscillator(DSP.SAW, frequency, 1.0, bufferSize, sampleRate);
+var dft = new dsp.DFT(bufferSize, sampleRate);
+var osc = new dsp.Oscillator(dsp.DSP.SAW, frequency, 1.0, bufferSize, sampleRate);
 var signal = osc.generate();
 
 var duration = benchmark(function() { dft.forward(signal); }, 20);
@@ -16,5 +20,5 @@ for (var i = 0; i < dft.spectrum.length; i++) {
 
 var peakFreq = dft.getBandFrequency(dft.peakBand);
 
-print("Detected peak: " + peakFreq + " Hz (error " + Math.abs(peakFreq - frequency) + " Hz)");
-print("20 DFTs: " + (duration) + " ms (" + ((duration) / 20) + "ms per DFT)\n");
+console.log("Detected peak: " + peakFreq + " Hz (error " + Math.abs(peakFreq - frequency) + " Hz)");
+console.log("20 DFTs: " + (duration) + " ms (" + ((duration) / 20) + "ms per DFT)\n");
