+++ bench/index.js
@@ -1,0 +1,6 @@
+/* eslint-disable */
+FREQUENCY = 440;
+require("./deinterleave");
+require("./dft");
+require("./fft");
+require("./rfft");
