+++ examples/biquad.html
@@ -11,7 +11,7 @@
     <script language="javascript" src="js/init.js"></script>
 
     <!-- Load DSP.js -->
-    <script src="../dsp.js"></script>
+    <script src="../dist/dsp.js"></script>
     <script>
       $(function() {
         $('#freq').slider({ orientation: 'vertical', range: 'min', min: 60, max: 10000, step: 1, value: 880, slide: changeF0 });
