+++ examples/vocoder.html
@@ -11,7 +11,7 @@
     <script language="javascript" src="js/init.js"></script>
 
     <!-- Load DSP.js -->
-    <script src="../dsp.js"></script>
+    <script src="../dist/dsp.js"></script>
     <script>
     </script>
