+++ lib/adsr.js
@@ -1,0 +1,134 @@
+/**
+ * ADSR Envelope
+ *
+ * @name ADSR
+ * @param {Number} attack The attack length in seconds
+ * @param {Number} decay The decay length in seconds
+ * @param {Number} sustain The sustain level
+ * @param {Number} release The release length in seconds
+ * @param {Number} sampleRate The the sample rate
+ *
+ * @constructor
+ */
+function ADSR(attackLength, decayLength, sustainLevel, sustainLength, releaseLength, sampleRate) {
+  this.sampleRate = sampleRate;
+  // Length in seconds
+  this.attackLength  = attackLength;
+  this.decayLength   = decayLength;
+  this.sustainLevel  = sustainLevel;
+  this.sustainLength = sustainLength;
+  this.releaseLength = releaseLength;
+  this.sampleRate    = sampleRate;
+
+  // Length in samples
+  this.attackSamples  = attackLength  * sampleRate;
+  this.decaySamples   = decayLength   * sampleRate;
+  this.sustainSamples = sustainLength * sampleRate;
+  this.releaseSamples = releaseLength * sampleRate;
+
+  // Updates the envelope sample positions
+  this.update = function() {
+    this.attack         =                this.attackSamples;
+    this.decay          = this.attack  + this.decaySamples;
+    this.sustain        = this.decay   + this.sustainSamples;
+    this.release        = this.sustain + this.releaseSamples;
+  };
+
+  this.update();
+
+  this.samplesProcessed = 0;
+}
+
+/**
+ * Start the envelope
+ */
+ADSR.prototype.noteOn = function() {
+  this.samplesProcessed = 0;
+  this.sustainSamples = this.sustainLength * this.sampleRate;
+  this.update();
+};
+
+/**
+ * Stop the envelope
+ *
+ * Send a note off when using a sustain of infinity to let the envelope enter
+ * the release phase
+ */
+ADSR.prototype.noteOff = function() {
+  this.sustainSamples = this.samplesProcessed - this.decaySamples;
+  this.update();
+};
+
+/**
+ * Process sample
+ */
+ADSR.prototype.processSample = function(sample) {
+  var amplitude = 0;
+
+  if ( this.samplesProcessed <= this.attack ) {
+    amplitude = 0 + (1 - 0) * ((this.samplesProcessed - 0) / (this.attack - 0));
+  } else if ( this.samplesProcessed > this.attack && this.samplesProcessed <= this.decay ) {
+    amplitude = 1 + (this.sustainLevel - 1) * ((this.samplesProcessed - this.attack) / (this.decay - this.attack));
+  } else if ( this.samplesProcessed > this.decay && this.samplesProcessed <= this.sustain ) {
+    amplitude = this.sustainLevel;
+  } else if ( this.samplesProcessed > this.sustain && this.samplesProcessed <= this.release ) {
+    amplitude = this.sustainLevel + (0 - this.sustainLevel) * ((this.samplesProcessed - this.sustain) / (this.release - this.sustain));
+  }
+
+  return sample * amplitude;
+};
+
+/**
+ * Get current value
+ * @return {Number} amplitude
+ */
+ADSR.prototype.value = function() {
+  var amplitude = 0;
+
+  if ( this.samplesProcessed <= this.attack ) {
+    amplitude = 0 + (1 - 0) * ((this.samplesProcessed - 0) / (this.attack - 0));
+  } else if ( this.samplesProcessed > this.attack && this.samplesProcessed <= this.decay ) {
+    amplitude = 1 + (this.sustainLevel - 1) * ((this.samplesProcessed - this.attack) / (this.decay - this.attack));
+  } else if ( this.samplesProcessed > this.decay && this.samplesProcessed <= this.sustain ) {
+    amplitude = this.sustainLevel;
+  } else if ( this.samplesProcessed > this.sustain && this.samplesProcessed <= this.release ) {
+    amplitude = this.sustainLevel + (0 - this.sustainLevel) * ((this.samplesProcessed - this.sustain) / (this.release - this.sustain));
+  }
+
+  return amplitude;
+};
+
+/**
+ * Process a buffer
+ * @param {Array} buffer
+ */
+ADSR.prototype.process = function(buffer) {
+  for ( var i = 0; i < buffer.length; i++ ) {
+    buffer[i] *= this.value();
+
+    this.samplesProcessed++;
+  }
+
+  return buffer;
+};
+
+/**
+ * Test if the envelope is active
+ * @return {Boolean}
+ */
+ADSR.prototype.isActive = function() {
+  if ( this.samplesProcessed > this.release || this.samplesProcessed === -1 ) {
+    return false;
+  } else {
+    return true;
+  }
+};
+
+/**
+ * Disable the envelope
+ */
+ADSR.prototype.disable = function() {
+  this.samplesProcessed = -1;
+};
+
+module.exports = ADSR;
