+++ lib/biquad.js
@@ -1,0 +1,288 @@
+/* global Float64Array */
+var DSP = require("./dsp");
+var sinh = require("./sinh");
+
+/**
+ * Biquad filter
+ *
+ * Created by Ricard Marxer <email@ricardmarxer.com> on 2010-05-23.
+ *  Copyright 2010 Ricard Marxer. All rights reserved.
+ *
+ * Implementation based on:
+ * http://www.musicdsp.org/files/Audio-EQ-Cookbook.txt
+ *
+ * @param {Number} type
+ * @param {Number} sampleRate
+ * @constructor
+ */
+function Biquad(type, sampleRate) {
+  this.Fs = sampleRate;
+  this.type = type;  // type of the filter
+  this.parameterType = DSP.Q; // type of the parameter
+
+  this.x_1_l = 0;
+  this.x_2_l = 0;
+  this.y_1_l = 0;
+  this.y_2_l = 0;
+
+  this.x_1_r = 0;
+  this.x_2_r = 0;
+  this.y_1_r = 0;
+  this.y_2_r = 0;
+
+  this.b0 = 1;
+  this.a0 = 1;
+
+  this.b1 = 0;
+  this.a1 = 0;
+
+  this.b2 = 0;
+  this.a2 = 0;
+
+  this.b0a0 = this.b0 / this.a0;
+  this.b1a0 = this.b1 / this.a0;
+  this.b2a0 = this.b2 / this.a0;
+  this.a1a0 = this.a1 / this.a0;
+  this.a2a0 = this.a2 / this.a0;
+
+  this.f0 = 3000;   // "wherever it"s happenin", man."  Center Frequency or
+                    // Corner Frequency, or shelf midpoint frequency, depending
+                    // on which filter type.  The "significant frequency".
+
+  this.dBgain = 12; // used only for peaking and shelving filters
+
+  this.Q = 1;       // the EE kind of definition, except for peakingEQ in which A*Q is
+                    // the classic EE Q.  That adjustment in definition was made so that
+                    // a boost of N dB followed by a cut of N dB for identical Q and
+                    // f0/Fs results in a precisely flat unity gain filter or "wire".
+
+  this.BW = -3;     // the bandwidth in octaves (between -3 dB frequencies for BPF
+                    // and notch or between midpoint (dBgain/2) gain frequencies for
+                    // peaking EQ
+
+  this.S = 1;       // a "shelf slope" parameter (for shelving EQ only).  When S = 1,
+                    // the shelf slope is as steep as it can be and remain monotonically
+                    // increasing or decreasing gain with frequency.  The shelf slope, in
+                    // dB/octave, remains proportional to S for all other values for a
+                    // fixed f0/Fs and dBgain.
+
+  this.coefficients = function() {
+    var b = [this.b0, this.b1, this.b2];
+    var a = [this.a0, this.a1, this.a2];
+    return {b: b, a:a};
+  };
+
+  this.setFilterType = function(type) {
+    this.type = type;
+    this.recalculateCoefficients();
+  };
+
+  this.setSampleRate = function(rate) {
+    this.Fs = rate;
+    this.recalculateCoefficients();
+  };
+
+  this.setQ = function(q) {
+    this.parameterType = DSP.Q;
+    this.Q = Math.max(Math.min(q, 115.0), 0.001);
+    this.recalculateCoefficients();
+  };
+
+  this.setBW = function(bw) {
+    this.parameterType = DSP.BW;
+    this.BW = bw;
+    this.recalculateCoefficients();
+  };
+
+  this.setS = function(s) {
+    this.parameterType = DSP.S;
+    this.S = Math.max(Math.min(s, 5.0), 0.0001);
+    this.recalculateCoefficients();
+  };
+
+  this.setF0 = function(freq) {
+    this.f0 = freq;
+    this.recalculateCoefficients();
+  };
+
+  this.setDbGain = function(g) {
+    this.dBgain = g;
+    this.recalculateCoefficients();
+  };
+
+  this.recalculateCoefficients = function() {
+    var A;
+    if (type === DSP.PEAKING_EQ || type === DSP.LOW_SHELF || type === DSP.HIGH_SHELF ) {
+      A = Math.pow(10, (this.dBgain/40));  // for peaking and shelving EQ filters only
+    } else {
+      A  = Math.sqrt( Math.pow(10, (this.dBgain/20)) );
+    }
+
+    var w0 = DSP.TWO_PI * this.f0 / this.Fs;
+
+    var cosw0 = Math.cos(w0);
+    var sinw0 = Math.sin(w0);
+
+    var alpha = 0;
+
+    switch (this.parameterType) {
+      case DSP.Q:
+        alpha = sinw0/(2*this.Q);
+        break;
+
+      case DSP.BW:
+        alpha = sinw0 * sinh( Math.LN2/2 * this.BW * w0/sinw0 );
+        break;
+
+      case DSP.S:
+        alpha = sinw0/2 * Math.sqrt( (A + 1/A)*(1/this.S - 1) + 2 );
+        break;
+    }
+
+    /**
+        FYI: The relationship between bandwidth and Q is
+             1/Q = 2*sinh(ln(2)/2*BW*w0/sin(w0))     (digital filter w BLT)
+        or   1/Q = 2*sinh(ln(2)/2*BW)             (analog filter prototype)
+
+        The relationship between shelf slope and Q is
+             1/Q = sqrt((A + 1/A)*(1/S - 1) + 2)
+    */
+
+    var coeff;
+
+    switch (this.type) {
+      case DSP.LPF:       // H(s) = 1 / (s^2 + s/Q + 1)
+        this.b0 =  (1 - cosw0)/2;
+        this.b1 =   1 - cosw0;
+        this.b2 =  (1 - cosw0)/2;
+        this.a0 =   1 + alpha;
+        this.a1 =  -2 * cosw0;
+        this.a2 =   1 - alpha;
+        break;
+
+      case DSP.HPF:       // H(s) = s^2 / (s^2 + s/Q + 1)
+        this.b0 =  (1 + cosw0)/2;
+        this.b1 = -(1 + cosw0);
+        this.b2 =  (1 + cosw0)/2;
+        this.a0 =   1 + alpha;
+        this.a1 =  -2 * cosw0;
+        this.a2 =   1 - alpha;
+        break;
+
+      case DSP.BPF_CONSTANT_SKIRT:       // H(s) = s / (s^2 + s/Q + 1)  (constant skirt gain, peak gain = Q)
+        this.b0 =   sinw0/2;
+        this.b1 =   0;
+        this.b2 =  -sinw0/2;
+        this.a0 =   1 + alpha;
+        this.a1 =  -2*cosw0;
+        this.a2 =   1 - alpha;
+        break;
+
+      case DSP.BPF_CONSTANT_PEAK:       // H(s) = (s/Q) / (s^2 + s/Q + 1)      (constant 0 dB peak gain)
+        this.b0 =   alpha;
+        this.b1 =   0;
+        this.b2 =  -alpha;
+        this.a0 =   1 + alpha;
+        this.a1 =  -2*cosw0;
+        this.a2 =   1 - alpha;
+        break;
+
+      case DSP.NOTCH:     // H(s) = (s^2 + 1) / (s^2 + s/Q + 1)
+        this.b0 =   1;
+        this.b1 =  -2*cosw0;
+        this.b2 =   1;
+        this.a0 =   1 + alpha;
+        this.a1 =  -2*cosw0;
+        this.a2 =   1 - alpha;
+        break;
+
+      case DSP.APF:       // H(s) = (s^2 - s/Q + 1) / (s^2 + s/Q + 1)
+        this.b0 =   1 - alpha;
+        this.b1 =  -2*cosw0;
+        this.b2 =   1 + alpha;
+        this.a0 =   1 + alpha;
+        this.a1 =  -2*cosw0;
+        this.a2 =   1 - alpha;
+        break;
+
+      case DSP.PEAKING_EQ:  // H(s) = (s^2 + s*(A/Q) + 1) / (s^2 + s/(A*Q) + 1)
+        this.b0 =   1 + alpha*A;
+        this.b1 =  -2*cosw0;
+        this.b2 =   1 - alpha*A;
+        this.a0 =   1 + alpha/A;
+        this.a1 =  -2*cosw0;
+        this.a2 =   1 - alpha/A;
+        break;
+
+      case DSP.LOW_SHELF:   // H(s) = A * (s^2 + (sqrt(A)/Q)*s + A)/(A*s^2 + (sqrt(A)/Q)*s + 1)
+        coeff = sinw0 * Math.sqrt( (A^2 + 1)*(1/this.S - 1) + 2*A );
+        this.b0 =    A*((A+1) - (A-1)*cosw0 + coeff);
+        this.b1 =  2*A*((A-1) - (A+1)*cosw0);
+        this.b2 =    A*((A+1) - (A-1)*cosw0 - coeff);
+        this.a0 =       (A+1) + (A-1)*cosw0 + coeff;
+        this.a1 =   -2*((A-1) + (A+1)*cosw0);
+        this.a2 =       (A+1) + (A-1)*cosw0 - coeff;
+        break;
+
+      case DSP.HIGH_SHELF:   // H(s) = A * (A*s^2 + (sqrt(A)/Q)*s + 1)/(s^2 + (sqrt(A)/Q)*s + A)
+        coeff = sinw0 * Math.sqrt( (A^2 + 1)*(1/this.S - 1) + 2*A );
+        this.b0 =    A*((A+1) + (A-1)*cosw0 + coeff);
+        this.b1 = -2*A*((A-1) + (A+1)*cosw0);
+        this.b2 =    A*((A+1) + (A-1)*cosw0 - coeff);
+        this.a0 =       (A+1) - (A-1)*cosw0 + coeff;
+        this.a1 =    2*((A-1) - (A+1)*cosw0);
+        this.a2 =       (A+1) - (A-1)*cosw0 - coeff;
+        break;
+    }
+
+    this.b0a0 = this.b0/this.a0;
+    this.b1a0 = this.b1/this.a0;
+    this.b2a0 = this.b2/this.a0;
+    this.a1a0 = this.a1/this.a0;
+    this.a2a0 = this.a2/this.a0;
+  };
+
+  this.process = function(buffer) {
+      //y[n] = (b0/a0)*x[n] + (b1/a0)*x[n-1] + (b2/a0)*x[n-2]
+      //       - (a1/a0)*y[n-1] - (a2/a0)*y[n-2]
+
+      var len = buffer.length;
+      var output = new Float64Array(len);
+
+      for ( var i=0; i<buffer.length; i++ ) {
+        output[i] = this.b0a0*buffer[i] + this.b1a0*this.x_1_l + this.b2a0*this.x_2_l - this.a1a0*this.y_1_l - this.a2a0*this.y_2_l;
+        this.y_2_l = this.y_1_l;
+        this.y_1_l = output[i];
+        this.x_2_l = this.x_1_l;
+        this.x_1_l = buffer[i];
+      }
+
+      return output;
+  };
+
+  this.processStereo = function(buffer) {
+      //y[n] = (b0/a0)*x[n] + (b1/a0)*x[n-1] + (b2/a0)*x[n-2]
+      //       - (a1/a0)*y[n-1] - (a2/a0)*y[n-2]
+
+      var len = buffer.length;
+      var output = new Float64Array(len);
+
+      for (var i = 0; i < len/2; i++) {
+        output[2*i] = this.b0a0*buffer[2*i] + this.b1a0*this.x_1_l + this.b2a0*this.x_2_l - this.a1a0*this.y_1_l - this.a2a0*this.y_2_l;
+        this.y_2_l = this.y_1_l;
+        this.y_1_l = output[2*i];
+        this.x_2_l = this.x_1_l;
+        this.x_1_l = buffer[2*i];
+
+        output[2*i+1] = this.b0a0*buffer[2*i+1] + this.b1a0*this.x_1_r + this.b2a0*this.x_2_r - this.a1a0*this.y_1_r - this.a2a0*this.y_2_r;
+        this.y_2_r = this.y_1_r;
+        this.y_1_r = output[2*i+1];
+        this.x_2_r = this.x_1_r;
+        this.x_1_r = buffer[2*i+1];
+      }
+
+      return output;
+  };
+}
+
+module.exports = Biquad;
