+++ lib/dft.js
@@ -1,0 +1,57 @@
+/* global Float64Array */
+var FourierTransform = require("./fourier");
+
+/**
+ * DFT is a class for calculating the Discrete Fourier Transform of a signal.
+ *
+ * @param {Number} bufferSize The size of the sample buffer to be computed
+ * @param {Number} sampleRate The sampleRate of the buffer (eg. 44100)
+ *
+ * @constructor
+ */
+function DFT(bufferSize, sampleRate) {
+  FourierTransform.call(this, bufferSize, sampleRate);
+
+  var N = bufferSize/2 * bufferSize;
+  var TWO_PI = 2 * Math.PI;
+
+  this.sinTable = new Float64Array(N);
+  this.cosTable = new Float64Array(N);
+
+  for (var i = 0; i < N; i++) {
+    this.sinTable[i] = Math.sin(i * TWO_PI / bufferSize);
+    this.cosTable[i] = Math.cos(i * TWO_PI / bufferSize);
+  }
+}
+
+/**
+ * Performs a forward transform on the sample buffer.
+ * Converts a time domain signal to frequency domain spectra.
+ *
+ * @param {Array} buffer The sample buffer
+ *
+ * @returns The frequency spectrum array
+ */
+DFT.prototype.forward = function(buffer) {
+  var real = this.real,
+    imag = this.imag,
+    rval,
+    ival;
+
+  for (var k = 0; k < this.bufferSize/2; k++) {
+    rval = 0.0;
+    ival = 0.0;
+
+    for (var n = 0; n < buffer.length; n++) {
+      rval += this.cosTable[k*n] * buffer[n];
+      ival += this.sinTable[k*n] * buffer[n];
+    }
+
+    real[k] = rval;
+    imag[k] = ival;
+  }
+
+  return this.calculateSpectrum();
+};
+
+module.exports = DFT;
