+++ lib/dsp.js
@@ -1,0 +1,315 @@
+/* global Float64Array */
+
+////////////////////////////////////////////////////////////////////////////////
+//                                  CONSTANTS                                 //
+////////////////////////////////////////////////////////////////////////////////
+
+/**
+ * DSP is an object which contains general purpose utility functions and constants
+ */
+var DSP = {
+  // Channels
+  LEFT:           0,
+  RIGHT:          1,
+  MIX:            2,
+
+  // Waveforms
+  SINE:           1,
+  TRIANGLE:       2,
+  SAW:            3,
+  SQUARE:         4,
+
+  // Filters
+  LOWPASS:        0,
+  HIGHPASS:       1,
+  BANDPASS:       2,
+  NOTCH:          3,
+
+  // Window functions
+  BARTLETT:       1,
+  BARTLETTHANN:   2,
+  BLACKMAN:       3,
+  COSINE:         4,
+  GAUSS:          5,
+  HAMMING:        6,
+  HANN:           7,
+  LANCZOS:        8,
+  RECTANGULAR:    9,
+  TRIANGULAR:     10,
+
+  // Loop modes
+  OFF:            0,
+  FW:             1,
+  BW:             2,
+  FWBW:           3,
+
+  // Math
+  TWO_PI:         2*Math.PI
+};
+
+// Setup arrays for platforms which do not support byte arrays
+function setupTypedArray(name, fallback) {
+  // check if TypedArray exists
+  // typeof on Minefield and Chrome return function, typeof on Webkit returns object.
+  if (typeof this[name] !== "function" && typeof this[name] !== "object") {
+    // nope.. check if WebGLArray exists
+    if (typeof this[fallback] === "function" && typeof this[fallback] !== "object") {
+      this[name] = this[fallback];
+    } else {
+      // nope.. set as Native JS array
+      this[name] = function(obj) {
+        if (obj instanceof Array) {
+          return obj;
+        } else if (typeof obj === "number") {
+          return new Array(obj);
+        }
+      };
+    }
+  }
+}
+
+setupTypedArray("Float64Array", "WebGLFloatArray");
+setupTypedArray("Int32Array",   "WebGLIntArray");
+setupTypedArray("Uint16Array",  "WebGLUnsignedShortArray");
+setupTypedArray("Uint8Array",   "WebGLUnsignedByteArray");
+
+
+////////////////////////////////////////////////////////////////////////////////
+//                            DSP UTILITY FUNCTIONS                           //
+////////////////////////////////////////////////////////////////////////////////
+
+/**
+ * Inverts the phase of a signal
+ *
+ * @param {Array} buffer A sample buffer
+ *
+ * @returns The inverted sample buffer
+ */
+DSP.invert = function(buffer) {
+  for (var i = 0, len = buffer.length; i < len; i++) {
+    buffer[i] *= -1;
+  }
+
+  return buffer;
+};
+
+/**
+ * Converts split-stereo (dual mono) sample buffers into a stereo interleaved sample buffer
+ *
+ * @param {Array} left  A sample buffer
+ * @param {Array} right A sample buffer
+ *
+ * @returns The stereo interleaved buffer
+ */
+DSP.interleave = function(left, right) {
+  if (left.length !== right.length) {
+    throw "Can not interleave. Channel lengths differ.";
+  }
+
+  var stereoInterleaved = new Float64Array(left.length * 2);
+
+  for (var i = 0, len = left.length; i < len; i++) {
+    stereoInterleaved[2*i]   = left[i];
+    stereoInterleaved[2*i+1] = right[i];
+  }
+
+  return stereoInterleaved;
+};
+
+/**
+ * Converts a stereo-interleaved sample buffer into split-stereo (dual mono) sample buffers
+ *
+ * @param {Array} buffer A stereo-interleaved sample buffer
+ *
+ * @returns an Array containing left and right channels
+ */
+DSP.deinterleave = (function() {
+  var left, right, mix, deinterleaveChannel = [];
+
+  deinterleaveChannel[DSP.MIX] = function(buffer) {
+    for (var i = 0, len = buffer.length/2; i < len; i++) {
+      mix[i] = (buffer[2*i] + buffer[2*i+1]) / 2;
+    }
+    return mix;
+  };
+
+  deinterleaveChannel[DSP.LEFT] = function(buffer) {
+    for (var i = 0, len = buffer.length/2; i < len; i++) {
+      left[i]  = buffer[2*i];
+    }
+    return left;
+  };
+
+  deinterleaveChannel[DSP.RIGHT] = function(buffer) {
+    for (var i = 0, len = buffer.length/2; i < len; i++) {
+      right[i]  = buffer[2*i+1];
+    }
+    return right;
+  };
+
+  return function(channel, buffer) {
+    left  = left  || new Float64Array(buffer.length/2);
+    right = right || new Float64Array(buffer.length/2);
+    mix   = mix   || new Float64Array(buffer.length/2);
+
+    if (buffer.length/2 !== left.length) {
+      left  = new Float64Array(buffer.length/2);
+      right = new Float64Array(buffer.length/2);
+      mix   = new Float64Array(buffer.length/2);
+    }
+
+    return deinterleaveChannel[channel](buffer);
+  };
+}());
+
+/**
+ * Separates a channel from a stereo-interleaved sample buffer
+ *
+ * @param {Array}  buffer A stereo-interleaved sample buffer
+ * @param {Number} channel A channel constant (LEFT, RIGHT, MIX)
+ *
+ * @returns an Array containing a signal mono sample buffer
+ */
+DSP.getChannel = DSP.deinterleave;
+
+/**
+ * Helper method (for Reverb) to mix two (interleaved) samplebuffers. It's possible
+ * to negate the second buffer while mixing and to perform a volume correction
+ * on the final signal.
+ *
+ * @param {Array} sampleBuffer1 Array containing Float values or a Float64Array
+ * @param {Array} sampleBuffer2 Array containing Float values or a Float64Array
+ * @param {Boolean} negate When true inverts/flips the audio signal
+ * @param {Number} volumeCorrection When you add multiple sample buffers, use this to tame your signal ;)
+ *
+ * @returns A new Float64Array interleaved buffer.
+ */
+DSP.mixSampleBuffers = function(sampleBuffer1, sampleBuffer2, negate, volumeCorrection){
+  var outputSamples = new Float64Array(sampleBuffer1);
+
+  for(var i = 0; i<sampleBuffer1.length; i++){
+    outputSamples[i] += (negate ? -sampleBuffer2[i] : sampleBuffer2[i]) / volumeCorrection;
+  }
+
+  return outputSamples;
+};
+
+// Biquad filter types
+DSP.LPF = 0;                // H(s) = 1 / (s^2 + s/Q + 1)
+DSP.HPF = 1;                // H(s) = s^2 / (s^2 + s/Q + 1)
+DSP.BPF_CONSTANT_SKIRT = 2; // H(s) = s / (s^2 + s/Q + 1)  (constant skirt gain, peak gain = Q)
+DSP.BPF_CONSTANT_PEAK = 3;  // H(s) = (s/Q) / (s^2 + s/Q + 1)      (constant 0 dB peak gain)
+DSP.NOTCH = 4;              // H(s) = (s^2 + 1) / (s^2 + s/Q + 1)
+DSP.APF = 5;                // H(s) = (s^2 - s/Q + 1) / (s^2 + s/Q + 1)
+DSP.PEAKING_EQ = 6;         // H(s) = (s^2 + s*(A/Q) + 1) / (s^2 + s/(A*Q) + 1)
+DSP.LOW_SHELF = 7;          // H(s) = A * (s^2 + (sqrt(A)/Q)*s + A)/(A*s^2 + (sqrt(A)/Q)*s + 1)
+DSP.HIGH_SHELF = 8;         // H(s) = A * (A*s^2 + (sqrt(A)/Q)*s + 1)/(s^2 + (sqrt(A)/Q)*s + A)
+
+// Biquad filter parameter types
+DSP.Q = 1;
+DSP.BW = 2; // SHARED with BACKWARDS LOOP MODE
+DSP.S = 3;
+
+/**
+ * Find RMS of signal
+ * @param {Array} buffer
+ */
+DSP.RMS = function(buffer) {
+  var total = 0;
+
+  for (var i = 0, n = buffer.length; i < n; i++) {
+    total += buffer[i] * buffer[i];
+  }
+
+  return Math.sqrt(total / n);
+};
+
+/**
+ * Find Peak of signal
+ * @param {Array} buffer
+ */
+DSP.Peak = function(buffer) {
+  var peak = 0;
+
+  for (var i = 0, n = buffer.length; i < n; i++) {
+    peak = (Math.abs(buffer[i]) > peak) ? Math.abs(buffer[i]) : peak;
+  }
+
+  return peak;
+};
+
+/**
+ * Magnitude to decibels
+ *
+ * Created by Ricard Marxer <email@ricardmarxer.com> on 2010-05-23.
+ * Copyright 2010 Ricard Marxer. All rights reserved.
+ *
+ * @param {Array} @buffer The array of magnitudes to convert to decibels
+ * @returns the array in decibels
+ *
+ */
+DSP.mag2db = function(buffer) {
+  var minDb = -120;
+  var minMag = Math.pow(10.0, minDb / 20.0);
+
+  var log = Math.log;
+  var max = Math.max;
+
+  var result = new Float64Array(buffer.length);
+  for (var i=0; i<buffer.length; i++) {
+    result[i] = 20.0*log(max(buffer[i], minMag));
+  }
+
+  return result;
+};
+
+/**
+ *  Frequency response
+ *
+ *  Created by Ricard Marxer <email@ricardmarxer.com> on 2010-05-23.
+ *  Copyright 2010 Ricard Marxer. All rights reserved.
+ *
+ *  Calculates the frequency response at the given points.
+ *
+ *  @param {Number} b The coefficients of the filter
+ *  @param {Number} a The coefficients of the filter
+ *  @param {Number} w The points (normally between -PI and PI) where to calculate the frequency response
+ *
+ *  @returns the frequency response in magnitude
+ */
+DSP.freqz = function(b, a, w) {
+  var i, j;
+
+  if (!w) {
+    w = new Float64Array(200);
+    for (i=0;i<w.length; i++) {
+      w[i] = DSP.TWO_PI/w.length * i - Math.PI;
+    }
+  }
+
+  var result = new Float64Array(w.length);
+
+  var sqrt = Math.sqrt;
+  var cos = Math.cos;
+  var sin = Math.sin;
+
+  for (i=0; i<w.length; i++) {
+    var numerator = {real:0.0, imag:0.0};
+    for (j=0; j<b.length; j++) {
+      numerator.real += b[j] * cos(-j*w[i]);
+      numerator.imag += b[j] * sin(-j*w[i]);
+    }
+
+    var denominator = {real:0.0, imag:0.0};
+    for (j=0; j<a.length; j++) {
+      denominator.real += a[j] * cos(-j*w[i]);
+      denominator.imag += a[j] * sin(-j*w[i]);
+    }
+
+    result[i] =  sqrt(numerator.real*numerator.real + numerator.imag*numerator.imag) / sqrt(denominator.real*denominator.real + denominator.imag*denominator.imag);
+  }
+
+  return result;
+};
+
+module.exports = DSP;
