+++ lib/iir-filter2.js
@@ -1,0 +1,88 @@
+/* global Float64Array */
+var ADSR = require("./adsr");
+
+/**
+ * IIRFilter2
+ *
+ * @constructor
+ * @param {Number} type
+ * @param {Number} cutoff
+ * @param {Number} resonance
+ * @param {Number} sampleRate
+ */
+function IIRFilter2(type, cutoff, resonance, sampleRate) {
+  this.type = type;
+  this.cutoff = cutoff;
+  this.resonance = resonance;
+  this.sampleRate = sampleRate;
+
+  this.f = new Float64Array(4);
+  this.f[0] = 0.0; // lp
+  this.f[1] = 0.0; // hp
+  this.f[2] = 0.0; // bp
+  this.f[3] = 0.0; // br
+
+  this.calcCoeff = function(cutoff, resonance) {
+    this.freq = 2 * Math.sin(Math.PI * Math.min(0.25, cutoff/(this.sampleRate*2)));
+    this.damp = Math.min(2 * (1 - Math.pow(resonance, 0.25)), Math.min(2, 2/this.freq - this.freq * 0.5));
+  };
+
+  this.calcCoeff(cutoff, resonance);
+}
+
+/**
+ * Process a buffer
+ * @param {Array} buffer
+ */
+IIRFilter2.prototype.process = function(buffer) {
+  var input, output;
+  var f = this.f;
+
+  for ( var i = 0; i < buffer.length; i++ ) {
+    input = buffer[i];
+
+    // first pass
+    f[3] = input - this.damp * f[2];
+    f[0] = f[0] + this.freq * f[2];
+    f[1] = f[3] - f[0];
+    f[2] = this.freq * f[1] + f[2];
+    output = 0.5 * f[this.type];
+
+    // second pass
+    f[3] = input - this.damp * f[2];
+    f[0] = f[0] + this.freq * f[2];
+    f[1] = f[3] - f[0];
+    f[2] = this.freq * f[1] + f[2];
+    output += 0.5 * f[this.type];
+
+    if (this.envelope) {
+      buffer[i] = (buffer[i] * (1 - this.envelope.value())) + (output * this.envelope.value());
+      this.envelope.samplesProcessed++;
+    } else {
+      buffer[i] = output;
+    }
+  }
+};
+
+/**
+ * Add an envelope to the filter
+ * @param {ADSR} envelope
+ */
+IIRFilter2.prototype.addEnvelope = function(envelope) {
+  if ( envelope instanceof ADSR ) {
+    this.envelope = envelope;
+  } else {
+    throw "This is not an envelope.";
+  }
+};
+
+/**
+ * Set filter parameters
+ * @param {Number} cutoff
+ * @param {Number} resonance
+ */
+IIRFilter2.prototype.set = function(cutoff, resonance) {
+  this.calcCoeff(cutoff, resonance);
+};
+
+module.exports = IIRFilter2;
