+++ lib/oscillator.js
@@ -1,0 +1,195 @@
+/* global Float64Array */
+var DSP = require("./dsp");
+
+/**
+ * Oscillator class for generating and modifying signals
+ *
+ * @param {Number} type       A waveform constant (eg. DSP.SINE)
+ * @param {Number} frequency  Initial frequency of the signal
+ * @param {Number} amplitude  Initial amplitude of the signal
+ * @param {Number} bufferSize Size of the sample buffer to generate
+ * @param {Number} sampleRate The sample rate of the signal
+ *
+ * @constructor
+ */
+function Oscillator(type, frequency, amplitude, bufferSize, sampleRate) {
+  this.frequency  = frequency;
+  this.amplitude  = amplitude;
+  this.bufferSize = bufferSize;
+  this.sampleRate = sampleRate;
+  //this.pulseWidth = pulseWidth;
+  this.frameCount = 0;
+
+  this.waveTableLength = 2048;
+
+  this.cyclesPerSample = frequency / sampleRate;
+
+  this.signal = new Float64Array(bufferSize);
+  this.envelope = null;
+
+  switch(parseInt(type, 10)) {
+    case DSP.TRIANGLE:
+      this.func = Oscillator.Triangle;
+      break;
+
+    case DSP.SAW:
+      this.func = Oscillator.Saw;
+      break;
+
+    case DSP.SQUARE:
+      this.func = Oscillator.Square;
+      break;
+
+    default:
+    case DSP.SINE:
+      this.func = Oscillator.Sine;
+      break;
+  }
+
+  this.generateWaveTable = function() {
+    Oscillator.waveTable[this.func] = new Float64Array(2048);
+    var waveTableTime = this.waveTableLength / this.sampleRate;
+    var waveTableHz = 1 / waveTableTime;
+
+    for (var i = 0; i < this.waveTableLength; i++) {
+      Oscillator.waveTable[this.func][i] = this.func(i * waveTableHz/this.sampleRate);
+    }
+  };
+
+  if ( typeof Oscillator.waveTable === "undefined" ) {
+    Oscillator.waveTable = {};
+  }
+
+  if ( typeof Oscillator.waveTable[this.func] === "undefined" ) {
+    this.generateWaveTable();
+  }
+
+  this.waveTable = Oscillator.waveTable[this.func];
+}
+
+/**
+ * Set the amplitude of the signal
+ *
+ * @param {Number} amplitude The amplitude of the signal (between 0 and 1)
+ */
+Oscillator.prototype.setAmp = function(amplitude) {
+  if (amplitude >= 0 && amplitude <= 1) {
+    this.amplitude = amplitude;
+  } else {
+    throw "Amplitude out of range (0..1).";
+  }
+};
+
+/**
+ * Set the frequency of the signal
+ *
+ * @param {Number} frequency The frequency of the signal
+ */
+Oscillator.prototype.setFreq = function(frequency) {
+  this.frequency = frequency;
+  this.cyclesPerSample = frequency / this.sampleRate;
+};
+
+/**
+ * Add an oscillator
+ * @param {Oscillator} oscillator The oscillator to be added to
+ * @return {Array} the current oscillator signal
+ */
+Oscillator.prototype.add = function(oscillator) {
+  for ( var i = 0; i < this.bufferSize; i++ ) {
+    //this.signal[i] += oscillator.valueAt(i);
+    this.signal[i] += oscillator.signal[i];
+  }
+
+  return this.signal;
+};
+
+/**
+ * Add a signal to the current generated osc signal
+ * @param {Array} signal
+ */
+Oscillator.prototype.addSignal = function(signal) {
+  for ( var i = 0; i < signal.length; i++ ) {
+    if ( i >= this.bufferSize ) {
+      break;
+    }
+    this.signal[i] += signal[i];
+
+    /*
+    // Constrain amplitude
+    if ( this.signal[i] > 1 ) {
+      this.signal[i] = 1;
+    } else if ( this.signal[i] < -1 ) {
+      this.signal[i] = -1;
+    }
+    */
+  }
+  return this.signal;
+};
+
+/**
+ * Add an envelope to the oscillator
+ * @param {ADSR} envelope
+ */
+Oscillator.prototype.addEnvelope = function(envelope) {
+  this.envelope = envelope;
+};
+
+/**
+ * Apply the oscillator envelope to its signal
+ */
+Oscillator.prototype.applyEnvelope = function() {
+  this.envelope.process(this.signal);
+};
+
+/**
+ * Get value
+ * @param {Number} offset
+ */
+Oscillator.prototype.valueAt = function(offset) {
+  return this.waveTable[offset % this.waveTableLength];
+};
+
+/**
+ * Generate the oscillator signal
+ * @return {Array} the signal
+ */
+Oscillator.prototype.generate = function() {
+  var frameOffset = this.frameCount * this.bufferSize;
+  var step = this.waveTableLength * this.frequency / this.sampleRate;
+  var offset;
+
+  for ( var i = 0; i < this.bufferSize; i++ ) {
+    //var step = (frameOffset + i) * this.cyclesPerSample % 1;
+    //this.signal[i] = this.func(step) * this.amplitude;
+    //this.signal[i] = this.valueAt(Math.round((frameOffset + i) * step)) * this.amplitude;
+    offset = Math.round((frameOffset + i) * step);
+    this.signal[i] = this.waveTable[offset % this.waveTableLength] * this.amplitude;
+  }
+
+  this.frameCount++;
+
+  return this.signal;
+};
+
+Oscillator.Sine = function(step) {
+  return Math.sin(DSP.TWO_PI * step);
+};
+
+Oscillator.Square = function(step) {
+  return step < 0.5 ? 1 : -1;
+};
+
+Oscillator.Saw = function(step) {
+  return 2 * (step - Math.round(step));
+};
+
+Oscillator.Triangle = function(step) {
+  return 1 - 4 * Math.abs(Math.round(step) - step);
+};
+
+Oscillator.Pulse = function() {
+  // stub
+};
+
+module.exports = Oscillator;
