+++ lib/reverb.js
@@ -1,0 +1,174 @@
+/* global Float64Array */
+var DSP = require("./dsp");
+var IIRFilter2 = require("./iir-filter2");
+var SingleDelay = require("./single-delay");
+var MultiDelay = require("./multi-delay");
+
+/**
+ * Reverb effect by Almer Thie (http://code.almeros.com).
+ * Copyright 2010 Almer Thie. All rights reserved.
+ * Example: http://code.almeros.com/code-examples/reverb-firefox-audio-api/
+ *
+ * This reverb consists of 6 SingleDelays, 6 MultiDelays and an IIRFilter2
+ * for each of the two stereo channels.
+ *
+ * Compatible with interleaved stereo buffers only!
+ *
+ * @param {Number} maxDelayInSamplesSize Maximum possible delay in samples (size of circular buffers)
+ * @param {Number} delayInSamples Initial delay in samples for internal (Single/Multi)delays
+ * @param {Number} masterVolume Initial master volume. Float value: 0.0 (silence), 1.0 (normal), >1.0 (amplify)
+ * @param {Number} mixVolume Initial reverb signal mix volume. Float value: 0.0 (silence), 1.0 (normal), >1.0 (amplify)
+ * @param {Number} delayVolume Initial feedback delay volume for internal (Single/Multi)delays. Float value: 0.0 (silence), 1.0 (normal), >1.0 (amplify)
+ * @param {Number} dampFrequency Initial low pass filter frequency. 0 to 44100 (depending on your maximum sampling frequency)
+ *
+ * @constructor
+ */
+function Reverb(maxDelayInSamplesSize, delayInSamples, masterVolume, mixVolume, delayVolume, dampFrequency) {
+  this.delayInSamples   = delayInSamples;
+  this.masterVolume     = masterVolume;
+  this.mixVolume       = mixVolume;
+  this.delayVolume     = delayVolume;
+  this.dampFrequency     = dampFrequency;
+
+  this.NR_OF_MULTIDELAYS = 6;
+  this.NR_OF_SINGLEDELAYS = 6;
+
+  this.LOWPASSL = new IIRFilter2(DSP.LOWPASS, dampFrequency, 0, 44100);
+  this.LOWPASSR = new IIRFilter2(DSP.LOWPASS, dampFrequency, 0, 44100);
+
+  this.singleDelays = [];
+
+  var i, delayMultiply;
+
+  for (i = 0; i < this.NR_OF_SINGLEDELAYS; i++) {
+    delayMultiply = 1.0 + (i/7.0); // 1.0, 1.1, 1.2...
+    this.singleDelays[i] = new SingleDelay(maxDelayInSamplesSize, Math.round(this.delayInSamples * delayMultiply), this.delayVolume);
+  }
+
+  this.multiDelays = [];
+
+  for (i = 0; i < this.NR_OF_MULTIDELAYS; i++) {
+    delayMultiply = 1.0 + (i/10.0); // 1.0, 1.1, 1.2...
+    this.multiDelays[i] = new MultiDelay(maxDelayInSamplesSize, Math.round(this.delayInSamples * delayMultiply), this.masterVolume, this.delayVolume);
+  }
+}
+
+/**
+ * Change the delay time in samples as a base for all delays.
+ *
+ * @param {Number} delayInSamples Delay in samples
+ */
+Reverb.prototype.setDelayInSamples = function (delayInSamples){
+  this.delayInSamples = delayInSamples;
+
+  var i, delayMultiply;
+
+  for (i = 0; i < this.NR_OF_SINGLEDELAYS; i++) {
+    delayMultiply = 1.0 + (i/7.0); // 1.0, 1.1, 1.2...
+    this.singleDelays[i].setDelayInSamples( Math.round(this.delayInSamples * delayMultiply) );
+  }
+
+  for (i = 0; i < this.NR_OF_MULTIDELAYS; i++) {
+    delayMultiply = 1.0 + (i/10.0); // 1.0, 1.1, 1.2...
+    this.multiDelays[i].setDelayInSamples( Math.round(this.delayInSamples * delayMultiply) );
+  }
+};
+
+/**
+ * Change the master volume.
+ *
+ * @param {Number} masterVolume Float value: 0.0 (silence), 1.0 (normal), >1.0 (amplify)
+ */
+Reverb.prototype.setMasterVolume = function (masterVolume){
+  this.masterVolume = masterVolume;
+};
+
+/**
+ * Change the reverb signal mix level.
+ *
+ * @param {Number} mixVolume Float value: 0.0 (silence), 1.0 (normal), >1.0 (amplify)
+ */
+Reverb.prototype.setMixVolume = function (mixVolume){
+  this.mixVolume = mixVolume;
+};
+
+/**
+ * Change all delays feedback volume.
+ *
+ * @param {Number} delayVolume Float value: 0.0 (silence), 1.0 (normal), >1.0 (amplify)
+ */
+Reverb.prototype.setDelayVolume = function (delayVolume){
+  this.delayVolume = delayVolume;
+
+  var i;
+
+  for (i = 0; i<this.NR_OF_SINGLEDELAYS; i++) {
+    this.singleDelays[i].setDelayVolume(this.delayVolume);
+  }
+
+  for (i = 0; i<this.NR_OF_MULTIDELAYS; i++) {
+    this.multiDelays[i].setDelayVolume(this.delayVolume);
+  }
+};
+
+/**
+ * Change the Low Pass filter frequency.
+ *
+ * @param {Number} dampFrequency low pass filter frequency. 0 to 44100 (depending on your maximum sampling frequency)
+ */
+Reverb.prototype.setDampFrequency = function (dampFrequency){
+  this.dampFrequency = dampFrequency;
+
+  this.LOWPASSL.set(dampFrequency, 0);
+  this.LOWPASSR.set(dampFrequency, 0);
+};
+
+/**
+ * Process a given interleaved float value Array and copies and adds the reverb signal.
+ *
+ * @param {Array} samples Array containing Float values or a Float64Array
+ *
+ * @returns A new Float64Array interleaved buffer.
+ */
+Reverb.prototype.process = function (interleavedSamples){
+  // NB. Make a copy to put in the output samples to return.
+  var outputSamples = new Float64Array(interleavedSamples.length);
+
+  // Perform low pass on the input samples to mimick damp
+  var leftRightMix = DSP.deinterleave(interleavedSamples);
+  this.LOWPASSL.process( leftRightMix[DSP.LEFT] );
+  this.LOWPASSR.process( leftRightMix[DSP.RIGHT] );
+  var filteredSamples = DSP.interleave(leftRightMix[DSP.LEFT], leftRightMix[DSP.RIGHT]);
+
+  var i;
+
+  // Process MultiDelays in parallel
+  for (i = 0; i<this.NR_OF_MULTIDELAYS; i++) {
+    // Invert the signal of every even multiDelay
+    outputSamples = DSP.mixSampleBuffers(outputSamples, this.multiDelays[i].process(filteredSamples), 2%i === 0, this.NR_OF_MULTIDELAYS);
+  }
+
+  // Process SingleDelays in series
+  var singleDelaySamples = new Float64Array(outputSamples.length);
+  for (i = 0; i<this.NR_OF_SINGLEDELAYS; i++) {
+    // Invert the signal of every even singleDelay
+    singleDelaySamples = DSP.mixSampleBuffers(singleDelaySamples, this.singleDelays[i].process(outputSamples), 2%i === 0, 1);
+  }
+
+  // Apply the volume of the reverb signal
+  for (i = 0; i<singleDelaySamples.length; i++) {
+    singleDelaySamples[i] *= this.mixVolume;
+  }
+
+  // Mix the original signal with the reverb signal
+  outputSamples = DSP.mixSampleBuffers(singleDelaySamples, interleavedSamples, 0, 1);
+
+  // Apply the master volume to the complete signal
+  for (i = 0; i<outputSamples.length; i++) {
+    outputSamples[i] *= this.masterVolume;
+  }
+
+  return outputSamples;
+};
+
+module.exports = Reverb;
