+++ lib/sampler.js
@@ -1,0 +1,133 @@
+/* global Float64Array */
+var DSP = require("./dsp");
+
+/**
+ * Sampler
+ *
+ * @constructor
+ * @param {} file
+ * @param {Number} bufferSize
+ * @param {Number} sampleRate
+ * @param {Number} playStart
+ * @param {Number} playEnd
+ * @param {Number} loopStart
+ * @param {Number} loopEnd
+ * @param {Number} loopMode
+ */
+function Sampler(file, bufferSize, sampleRate, playStart, playEnd, loopStart, loopEnd, loopMode) {
+  this.file = file;
+  this.bufferSize = bufferSize;
+  this.sampleRate = sampleRate;
+  this.playStart  = playStart || 0; // 0%
+  this.playEnd    = playEnd   || 1; // 100%
+  this.loopStart  = loopStart || 0;
+  this.loopEnd    = loopEnd   || 1;
+  this.loopMode   = loopMode  || DSP.OFF;
+  this.loaded     = false;
+  this.samples    = [];
+  this.signal     = new Float64Array(bufferSize);
+  this.frameCount = 0;
+  this.envelope   = null;
+  this.amplitude  = 1;
+  this.rootFrequency = 110; // A2 110
+  this.frequency  = 550;
+  this.step       = this.frequency / this.rootFrequency;
+  this.duration   = 0;
+  this.samplesProcessed = 0;
+  this.playhead   = 0;
+
+  var audio = /* new Audio();*/ document.createElement("AUDIO");
+  var self = this;
+
+  this.loadSamples = function(event) {
+    var buffer = DSP.getChannel(DSP.MIX, event.frameBuffer);
+    for ( var i = 0; i < buffer.length; i++) {
+      self.samples.push(buffer[i]);
+    }
+  };
+
+  this.loadComplete = function() {
+    // convert flexible js array into a fast typed array
+    self.samples = new Float64Array(self.samples);
+    self.loaded = true;
+  };
+
+  this.loadMetaData = function() {
+    self.duration = audio.duration;
+  };
+
+  audio.addEventListener("MozAudioAvailable", this.loadSamples, false);
+  audio.addEventListener("loadedmetadata", this.loadMetaData, false);
+  audio.addEventListener("ended", this.loadComplete, false);
+  audio.muted = true;
+  audio.src = file;
+  audio.play();
+}
+
+Sampler.prototype.applyEnvelope = function() {
+  this.envelope.process(this.signal);
+  return this.signal;
+};
+
+Sampler.prototype.generate = function() {
+  var loopWidth = this.playEnd * this.samples.length - this.playStart * this.samples.length;
+  var playStartSamples = this.playStart * this.samples.length; // ie 0.5 -> 50% of the length
+  var playEndSamples = this.playEnd * this.samples.length; // ie 0.5 -> 50% of the length
+
+  for ( var i = 0; i < this.bufferSize; i++ ) {
+    switch (this.loopMode) {
+      case DSP.OFF:
+        this.playhead = Math.round(this.samplesProcessed * this.step + playStartSamples);
+        if (this.playhead < (this.playEnd * this.samples.length) ) {
+          this.signal[i] = this.samples[this.playhead] * this.amplitude;
+        } else {
+          this.signal[i] = 0;
+        }
+        break;
+
+      case DSP.FW:
+        this.playhead = Math.round((this.samplesProcessed * this.step) % loopWidth + playStartSamples);
+        if (this.playhead < (this.playEnd * this.samples.length) ) {
+          this.signal[i] = this.samples[this.playhead] * this.amplitude;
+        }
+        break;
+
+      case DSP.BW:
+        this.playhead = playEndSamples - Math.round((this.samplesProcessed * this.step) % loopWidth);
+        if (this.playhead < (this.playEnd * this.samples.length) ) {
+          this.signal[i] = this.samples[this.playhead] * this.amplitude;
+        }
+        break;
+
+      case DSP.FWBW:
+        if ( Math.floor(this.samplesProcessed * this.step / loopWidth) % 2 === 0 ) {
+          this.playhead = Math.round((this.samplesProcessed * this.step) % loopWidth + playStartSamples);
+        } else {
+          this.playhead = playEndSamples - Math.round((this.samplesProcessed * this.step) % loopWidth);
+        }
+        if (this.playhead < (this.playEnd * this.samples.length) ) {
+          this.signal[i] = this.samples[this.playhead] * this.amplitude;
+        }
+        break;
+    }
+    this.samplesProcessed++;
+  }
+
+  this.frameCount++;
+
+  return this.signal;
+};
+
+Sampler.prototype.setFreq = function(frequency) {
+    var totalProcessed = this.samplesProcessed * this.step;
+    this.frequency = frequency;
+    this.step = this.frequency / this.rootFrequency;
+    this.samplesProcessed = Math.round(totalProcessed/this.step);
+};
+
+Sampler.prototype.reset = function() {
+  this.samplesProcessed = 0;
+  this.playhead = 0;
+};
+
+module.exports = Sampler;
