+++ package.json
@@ -2,7 +2,16 @@
   "name": "dsp.js",
   "version": "1.0.0",
   "description": "Digital Signal Processing for Javascript",
-  "main": "dsp.js",
+  "main": "lib/index.js",
+  "scripts": {
+    "prepublish": "npm run dist",
+    "predist": "npm run test && npm run lint && npm run docs",
+    "dist": "rm dist/* && browserify lib/index.js > dist/dsp.js && uglifyjs dist/dsp.js > dist/dsp.min.js",
+    "docs": "jsdoc2md \"lib/*.js\" > docs/API.md",
+    "test": "mocha test/*.js",
+    "lint": "eslint lib/*.js",
+    "bench": "node bench/index.js"
+  },
   "directories": {
     "example": "examples",
     "test": "test"
@@ -21,5 +30,15 @@
   "bugs": {
     "url": "https://github.com/corbanbrook/dsp.js/issues"
   },
-  "homepage": "https://github.com/corbanbrook/dsp.js#readme"
+  "homepage": "https://github.com/corbanbrook/dsp.js#readme",
+  "standard": {
+    "ignore": "**/*"
+  },
+  "devDependencies": {
+    "browserify": "^13.1.1",
+    "eslint": "^3.12.2",
+    "jsdoc-to-markdown": "^2.0.1",
+    "mocha": "^3",
+    "uglifyjs": "^2.4.10"
+  }
 }
