+++ test/audio-test/beatdetect-test.js
